requirements.txt
Flask==3.0.0
python-telegram-bot==20.7
requests==2.31.0
Pillow==10.1.0
gunicorn
